Rename and move everything to hopefully make it possible to use this package as a Golang package
