+++ server/freeze.go
@@ -2,23 +2,24 @@
 // The use of this source code is goverened by a BSD-style
 // license that can be found in the LICENSE-file.
 
-package main
+package server
 
 import (
 	"errors"
-	"github.com/golang/protobuf/proto"
 	"io"
 	"io/ioutil"
 	"log"
-	"mumble.info/grumble/pkg/acl"
-	"mumble.info/grumble/pkg/ban"
-	"mumble.info/grumble/pkg/freezer"
-	"mumble.info/grumble/pkg/mumbleproto"
-	"mumble.info/grumble/pkg/serverconf"
 	"os"
 	"path/filepath"
 	"strconv"
 	"time"
+
+	"github.com/digitalautonomy/grumble/pkg/acl"
+	"github.com/digitalautonomy/grumble/pkg/ban"
+	"github.com/digitalautonomy/grumble/pkg/freezer"
+	"github.com/digitalautonomy/grumble/pkg/mumbleproto"
+	"github.com/digitalautonomy/grumble/pkg/serverconf"
+	"github.com/golang/protobuf/proto"
 )
 
 // Freeze a server to disk and closes the log file.
