+++ server/freeze_windows.go
@@ -2,16 +2,16 @@
 // The use of this source code is goverened by a BSD-style
 // license that can be found in the LICENSE-file.
 
-// +build !windows
-
-package main
+package server
 
 import (
-	"github.com/golang/protobuf/proto"
 	"io/ioutil"
 	"os"
 	"path/filepath"
 	"strconv"
+
+	"github.com/digitalautonomy/grumble/pkg/replacefile"
+	"github.com/golang/protobuf/proto"
 )
 
 func (server *Server) freezeToFile() (err error) {
@@ -48,9 +48,20 @@ func (server *Server) freezeToFile() (err error) {
 	if err != nil {
 		return err
 	}
-	err = os.Rename(f.Name(), filepath.Join(Args.DataDir, "servers", strconv.FormatInt(server.Id, 10), "main.fz"))
-	if err != nil {
-		return err
+
+	src := f.Name()
+	dst := filepath.Join(Args.DataDir, "servers", strconv.FormatInt(server.Id, 10), "main.fz")
+	backup := filepath.Join(Args.DataDir, "servers", strconv.FormatInt(server.Id, 10), "backup.fz")
+
+	err = replacefile.ReplaceFile(dst, src, backup, replacefile.Flag(0))
+	// If the dst file does not exist (as in, on first launch)
+	// fall back to os.Rename. ReplaceFile does not work if the
+	// dst file is not there.
+	if os.IsNotExist(err) {
+		err = os.Rename(src, dst)
+		if err != nil {
+			return err
+		}
 	}
 
 	return nil
