+++ server/grumble.go
@@ -2,23 +2,24 @@
 // The use of this source code is goverened by a BSD-style
 // license that can be found in the LICENSE-file.
 
-package main
+package server
 
 import (
 	"flag"
 	"fmt"
 	"log"
-	"mumble.info/grumble/pkg/blobstore"
-	"mumble.info/grumble/pkg/logtarget"
 	"os"
 	"path/filepath"
 	"regexp"
+
+	"github.com/digitalautonomy/grumble/pkg/blobstore"
+	"github.com/digitalautonomy/grumble/pkg/logtarget"
 )
 
 var servers map[int64]*Server
 var blobStore blobstore.BlobStore
 
-func main() {
+func exampleMain() {
 	var err error
 
 	flag.Parse()
