+++ server/version.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 var (
 	version   = "1.0~devel"
