+++ server/voicetarget.go
@@ -2,9 +2,9 @@
 // The use of this source code is goverened by a BSD-style
 // license that can be found in the LICENSE-file.
 
-package main
+package server
 
-import "mumble.info/grumble/pkg/acl"
+import "github.com/digitalautonomy/grumble/pkg/acl"
 
 // A VoiceTarget holds information about a single
 // VoiceTarget entry of a Client.
